Refactor Explorer: Extract Upload/Download
