+++ .settings/settings.json
@@ -1,0 +1,10 @@
+// Place your settings in this file to overwrite default and user settings.
+{
+	"search.excludeFolders": [
+	    ".git",
+	    "node_modules",
+	    "bower_components",
+	    "dist",
+		"cache"
+	],
+}
